# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityOut": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled_by_default": {
            "title": "Enabled By Default",
            "type": "boolean"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          }
        },
        "required": [
          "code",
          "description",
          "risk",
          "enabled_by_default"
        ],
        "title": "CapabilityOut",
        "type": "object"
      },
...
      "FamilyCapabilityOut": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled_by_default": {
            "title": "Enabled By Default",
            "type": "boolean"
          },
          "fixture_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Ref"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "verified_by_fixture": {
            "title": "Verified By Fixture",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "description",
          "risk",
          "enabled_by_default",
          "status",
          "verified_by_fixture",
          "fixture_ref",
          "note"
        ],
        "title": "FamilyCapabilityOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OidProfileDetail": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/OidProfileEntryOut"
            },
            "title": "Entries",
            "type": "array"
          },
          "family_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Slug"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "vendor_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Slug"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "status",
          "description",
          "created_at",
          "activated_at",
          "vendor_slug",
          "family_slug",
          "entries"
        ],
        "title": "OidProfileDetail",
        "type": "object"
      },
...
      "OidProfileSummary": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "family_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Slug"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "vendor_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Slug"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "status",
          "description",
          "created_at",
          "activated_at",
          "vendor_slug",
          "family_slug",
          "entries"
        ],
        "title": "OidProfileSummary",
        "type": "object"
      },
...
      "PollingToggle": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "PollingToggle",
        "type": "object"
      },
...
      "VendorDetail": {
        "additionalProperties": false,
        "properties": {
          "discovery_oids": {
            "items": {
              "type": "string"
            },
            "title": "Discovery Oids",
            "type": "array"
          },
          "discovery_retries": {
            "title": "Discovery Retries",
            "type": "integer"
          },
          "discovery_timeout_ms": {
            "title": "Discovery Timeout Ms",
            "type": "integer"
          },
          "families": {
            "title": "Families",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_families": {
            "items": {
              "$ref": "#/components/schemas/ModelFamilyOut"
            },
            "title": "Model Families",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "polling_disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polling Disabled Reason"
          },
          "polling_enabled": {
            "title": "Polling Enabled",
            "type": "boolean"
          },
          "polling_toggled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polling Toggled At"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "notes",
          "discovery_oids",
          "discovery_timeout_ms",
          "discovery_retries",
          "polling_enabled",
          "polling_disabled_reason",
          "polling_toggled_at",
          "families",
          "model_families"
        ],
        "title": "VendorDetail",
        "type": "object"
      },
      "VendorOut": {
        "additionalProperties": false,
        "properties": {
          "discovery_oids": {
            "items": {
              "type": "string"
            },
            "title": "Discovery Oids",
            "type": "array"
          },
          "discovery_retries": {
            "title": "Discovery Retries",
            "type": "integer"
          },
          "discovery_timeout_ms": {
            "title": "Discovery Timeout Ms",
            "type": "integer"
          },
          "families": {
            "title": "Families",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "polling_disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polling Disabled Reason"
          },
          "polling_enabled": {
            "title": "Polling Enabled",
            "type": "boolean"
          },
          "polling_toggled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polling Toggled At"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "notes",
          "discovery_oids",
          "discovery_timeout_ms",
          "discovery_retries",
          "polling_enabled",
          "polling_disabled_reason",
          "polling_toggled_at",
          "families"
        ],
        "title": "VendorOut",
        "type": "object"
      },
      "VendorUpdate": {
        "properties": {
          "discovery_retries": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Retries"
          },
          "discovery_timeout_ms": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 200.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Timeout Ms"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "VendorUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CyberSathy-NMS",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "list_capabilities_api_v1_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CapabilityOut"
                  },
                  "title": "Response List Capabilities Api V1 Capabilities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "registries"
        ]
      }
    },
...
    "/api/v1/oid-profiles": {
      "get": {
        "operationId": "list_oid_profiles_api_v1_oid_profiles_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OidProfileSummary"
                  },
                  "title": "Response List Oid Profiles Api V1 Oid Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Oid Profiles",
        "tags": [
          "registries"
        ]
      }
    },
    "/api/v1/oid-profiles/{profile_id}": {
      "get": {
        "operationId": "get_oid_profile_api_v1_oid_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OidProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Oid Profile",
        "tags": [
          "registries"
        ]
      }
    },
...
    "/api/v1/vendors": {
      "get": {
        "operationId": "list_vendors_api_v1_vendors_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VendorOut"
                  },
                  "title": "Response List Vendors Api V1 Vendors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Vendors",
        "tags": [
          "registries"
        ]
      }
    },
    "/api/v1/vendors/{slug}": {
      "get": {
        "operationId": "get_vendor_api_v1_vendors__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vendor",
        "tags": [
          "registries"
        ]
      },
      "patch": {
        "operationId": "update_vendor_api_v1_vendors__slug__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Vendor",
        "tags": [
          "registries"
        ]
      }
    },
    "/api/v1/vendors/{slug}/polling": {
      "put": {
        "operationId": "toggle_vendor_polling_api_v1_vendors__slug__polling_put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Vendor Polling",
        "tags": [
          "registries"
        ]
      }
    },
    "/api/v1/vendors/{vendor_slug}/families/{family_slug}/capabilities": {
      "get": {
        "operationId": "family_capabilities_api_v1_vendors__vendor_slug__families__family_slug__capabilities_get",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_slug",
            "required": true,
            "schema": {
              "title": "Vendor Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "family_slug",
            "required": true,
            "schema": {
              "title": "Family Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FamilyCapabilityOut"
                  },
                  "title": "Response Family Capabilities Api V1 Vendors  Vendor Slug  Families  Family Slug  Capabilities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Family Capabilities",
        "tags": [
          "registries"
        ]
      }
    },
    "/api/v1/vendors/{vendor_slug}/families/{family_slug}/polling": {
      "put": {
        "operationId": "toggle_family_polling_api_v1_vendors__vendor_slug__families__family_slug__polling_put",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_slug",
            "required": true,
            "schema": {
              "title": "Vendor Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "family_slug",
            "required": true,
            "schema": {
              "title": "Family Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollingToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelFamilyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Family Polling",
        "tags": [
          "registries"
        ]
      }
    },
...
  }
}